+++ hansel/__init__.py
@@ -3,6 +3,7 @@ from .crumb  import Crumb
 from .mktree import mktree
 from .utils  import (ParameterGrid,
                      intersection,
+                     difference,
                      groupby_pattern,
                      crumb_copy,
                      crumb_link)