+++ hansel/cli/cli.py
@@ -4,6 +4,7 @@ import click
 from .utils import (CONTEXT_SETTINGS,
                     CrumbPath,
                     echo_list,
+                    _print_values_map_as_csv,
                     check_not_none,)
 
 
@@ -123,17 +124,61 @@ def link(src_crumb, dst_crumb, quiet, ignore, exist_ok):
 @click.option('-o', '--on', type=str, multiple=True,
               help='Argument name to check for intersection. You can use this '
                    'argument more than once.')
-@click.option('-a', '--arg', type=str,
-              help='Will not print full paths, but only values for this '
-                   'crumb argument.')
-def intersect(crumb1, crumb2, on, arg):
+@click.option('-b', '--base', type=click.Choice(['1', '2']), default='0',
+              help='1 or 2, to indicate `crumb1` or `crumb2` as a base crumb2 '
+                   'to print the results')
+def intersect(crumb1, crumb2, on, base):
     """Return the intersection between crumb1 and crumb2 on a given argument.
+
+    Will not print full paths, but only values for the crumb arguments in `on`.
+    Unless you specify the crumb you want to use as a base with the `base`
+    argument.
 
     Examples: \n
     crumb intersect --on "sid" "/data/hansel/cobre/{sid}/{session}/{img}" "/data/hansel/cobre2/{sid}/{img}"\n
     """
     from .. import intersection
 
-    intersects = intersection(crumb1, crumb2, on=on)
-    for values in intersects:
-        click.echo(', '.join([arg[1] for arg in values]))
+    values = intersection(crumb1, crumb2, on=on)
+    if base == '0':
+        _print_values_map_as_csv(values)
+    else:
+        if base == '1':
+            base_crumb = crumb1
+        else:
+            base_crumb = crumb2
+
+        echo_list(base_crumb.build_paths(values))
+
+
+@cli.command(context_settings=CONTEXT_SETTINGS)
+@click.argument('crumb1', type=CrumbPath(), callback=check_not_none)
+@click.argument('crumb2', type=CrumbPath(), callback=check_not_none)
+@click.option('-o', '--on', type=str, multiple=True,
+              help='Argument name to check for intersection. You can use this '
+                   'argument more than once.')
+@click.option('-b', '--base', type=click.Choice(['1', '2']), default='0',
+              help='1 or 2, to indicate `crumb1` or `crumb2` as a base crumb2 '
+                   'to print the results')
+def diff(crumb1, crumb2, on, base):
+    """Return the difference crumb1 - crumb2 on a given argument.
+
+    Will not print full paths, but only values for the crumb arguments in `on`.
+    Unless you specify the crumb you want to use as a base with the `base`
+    argument.
+
+    Examples: \n
+    crumb diff --on "sid" "/data/hansel/cobre/{sid}/{session}/{img}" "/data/hansel/cobre2/{sid}/{img}"\n
+    """
+    from .. import difference
+
+    values = difference(crumb1, crumb2, on=on)
+    if base == '0':
+        _print_values_map_as_csv(values)
+    else:
+        if base == '1':
+            base_crumb = crumb1
+        else:
+            base_crumb = crumb2
+
+        echo_list(base_crumb.build_paths(values))