+++ hansel/tests/test_utils.py
@@ -13,6 +13,7 @@ from hansel.utils import (rm_dups,
                           ParameterGrid,
                           list_intersection,
                           intersection,
+                          difference,
                           crumb_copy,
                           crumb_link,
                           _get_matching_items,
@@ -226,6 +227,75 @@ def test_intersection():
     pytest.raises(KeyError, intersection, tmp_crumb1, Crumb(op.expanduser('~/{files}')))
 
     pytest.raises(KeyError, intersection, tmp_crumb1, Crumb(op.expanduser('~/{files}')), on=['files'])
+
+
+def test_difference():
+
+    crumb = Crumb("{base_dir}/raw/{subject_id}/{session_id}/{modality}/{image}")
+    base_dir1 = tempfile.mkdtemp(prefix='crumbtest1_')
+    tmp_crumb1 = crumb.replace(base_dir=base_dir1)
+
+    base_dir2 = tempfile.mkdtemp(prefix='crumbtest2_')
+    tmp_crumb2 = crumb.replace(base_dir=base_dir2)
+
+    assert not op.exists(tmp_crumb1._path)
+    assert not op.exists(tmp_crumb2._path)
+
+    assert not tmp_crumb1.has_files()
+    assert not tmp_crumb2.has_files()
+
+    values_dict1 = {'session_id': ['session_{:02}'.format(i) for i in range(4)],
+                    'subject_id': ['subj_{:03}'.format(i)    for i in range(5)],
+                    'modality':   ['anat'],
+                    'image':      ['mprage1.nii', 'mprage2.nii', 'mprage3.nii'],
+                    }
+
+    values_dict2 = {'session_id': ['session_{:02}'.format(i) for i in range(2)],
+                    'subject_id': ['subj_{:03}'.format(i)    for i in range(3)],
+                    'modality':   ['anat'],
+                    'image':      ['mprage1.nii', 'mprage2.nii', 'mprage3.nii'],
+                    }
+
+    _ = mktree(tmp_crumb1, list(ParameterGrid(values_dict1)))
+    _ = mktree(tmp_crumb2, list(ParameterGrid(values_dict2)))
+
+    assert op.exists(tmp_crumb1.split()[0])
+    assert op.exists(tmp_crumb2.split()[0])
+
+    n_extra_subjs = len(values_dict1['subject_id']) - len(values_dict2['subject_id'])
+    assert difference(tmp_crumb1, tmp_crumb2, on=['subject_id']) == \
+           [(('subject_id', val), ) for val in values_dict1['subject_id'][-n_extra_subjs:]]
+
+    assert difference(tmp_crumb1, tmp_crumb2, on=['subject_id', 'modality']) == \
+                      [(('subject_id', 'subj_003'), ('modality', 'anat')),
+                       (('subject_id', 'subj_004'), ('modality', 'anat')),]
+
+    han_crumb = tmp_crumb2.replace(subject_id='hansel')
+    assert difference(tmp_crumb1, han_crumb, on=['subject_id']) == \
+        [(('subject_id', val), ) for val in values_dict1['subject_id']]
+
+    s0_crumb = tmp_crumb2.replace(subject_id='subj_000')
+    assert difference(tmp_crumb1, s0_crumb, on=['subject_id']) == \
+        [(('subject_id', val), ) for val in values_dict1['subject_id'][1:]]
+
+    assert difference(tmp_crumb1, s0_crumb, on=['subject_id', 'modality']) == \
+                    [(('subject_id', 'subj_001'), ('modality', 'anat')),
+                     (('subject_id', 'subj_002'), ('modality', 'anat')),
+                     (('subject_id', 'subj_003'), ('modality', 'anat')),
+                     (('subject_id', 'subj_004'), ('modality', 'anat')),]
+
+    # test raises
+    pytest.raises(KeyError, difference, tmp_crumb1, tmp_crumb2, on=['hansel'])
+
+    pytest.raises(KeyError, difference, tmp_crumb1, tmp_crumb2,
+                                        on=['subject_id', 'modality', 'hansel'])
+
+    pytest.raises(KeyError, difference, tmp_crumb1,
+                                        Crumb(op.expanduser('~/{files}')))
+
+    pytest.raises(KeyError, difference, tmp_crumb1,
+                                        Crumb(op.expanduser('~/{files}')),
+                                        on=['files'])
 
 
 def test_group_pattern():